requirements.txt
Flask==2.3.2
mysql-connector-python==8.1.0
python-dotenv==1.0.0
Flask-WTF==1.1.1
